Split `character_selection_ui` to its own crate.
This makes testing character selection widgets much easier as it
requires characters to be loaded. The `application_test_support` crate
depends on `character_selection` so we can't depend on that if the ui is
in that crate.

Issue #51

+++ crate/character_selection/src/character_selection.rs
@@ -6,5 +6,5 @@ pub enum CharacterSelection {
     #[derivative(Default)]
     Random,
     /// Character with a particular ID.
-    Id(u32),
+    Id(usize),
 }
